requirements.txt
playwright>=1.45.0
pandas>=2.2.0
python-dotenv==1.0.0
beautifulsoup4>=4.12.2